chore: fix clippy warnings and clean up debug output
- Remove debug eprintln! from VP8L encoder (huffman.rs, encode.rs)
- Add #[allow(dead_code)] to vp8l module for future infrastructure
- Fix unused variable warnings with underscore prefixes
- Fix clippy lints: manual_range_contains, manual_clamp, manual_div_ceil,
  iter_cloned_collect, manual_memcpy, let_and_return, needless_range_loop,
  unnecessary_map_or, get_first, unnecessary_cast, needless_borrows_for_generic_args
- Use .is_multiple_of() instead of % 2 != 0
- Use field_reassign_with_default pattern for Vp8lConfig initialization
- Use core::f64::consts::LN_2 instead of hardcoded value

+++ examples/analyze_mb.rs
@@ -85,13 +85,13 @@ fn main() {
                 .y_blocks
                 .iter()
                 .flat_map(|b| b.levels.iter())
-                .map(|&lv| lv.abs() as i32)
+                .map(|&lv| lv.abs())
                 .sum();
             let lib_sum: i32 = l
                 .y_blocks
                 .iter()
                 .flat_map(|b| b.levels.iter())
-                .map(|&lv| lv.abs() as i32)
+                .map(|&lv| lv.abs())
                 .sum();
 
             println!(
